+++ packages/docusaurus-plugin-content-blog/src/__tests__/blogFrontMatter.test.ts
@@ -5,10 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {
-  BlogPostFrontMatter,
-  validateBlogPostFrontMatter,
-} from '../blogFrontMatter';
+import {validateBlogPostFrontMatter} from '../blogFrontMatter';
+import type {BlogPostFrontMatter} from '@docusaurus/plugin-content-blog';
 import escapeStringRegexp from 'escape-string-regexp';
 
 // TODO this abstraction reduce verbosity but it makes it harder to debug
